BUG: Isolate unbias mechanics to the optimizers, not SINDy class
Previously, optimizers could not restrict the unbias argument, since it was
applied by SINDy,fit() to a wrapper object, not the optimizer itself.  This
change simplifies SINDy objects by letting all optimizers inherit the unbias
argument and _unbias method.  Those that wish to restrict it, can.  SINDy.fit()
no longer takes an unbias argument and no longer creates its own SINDyOptimizer
wrapper.

It also changes the name of SINDyOptimizer to WrappedOptimizer.
