+++ pysindy/optimizers/__init__.py
@@ -19,7 +19,7 @@ try:  # Waiting on PEP 690 to lazy import cvxpy
     from .stable_linear_sr3 import StableLinearSR3
 except ImportError:
     pass
-from .sindy_optimizer import SINDyOptimizer
+from .wrapped_optimizer import WrappedOptimizer
 from .sr3 import SR3
 from .ssr import SSR
 from .stlsq import STLSQ
