+++ pysindy/optimizers/base.py
@@ -62,6 +62,17 @@ class BaseOptimizer(LinearRegression, ComplexityMixin):
         Initial guess for coefficients ``coef_``.
         If None, the initial guess is obtained via a least-squares fit.
 
+    unbias:  Whether to perform an extra step of unregularized linear
+        regression to unbias the coefficients for the identified
+        support.  If an optimizer (``self.optimizer``) applies any type
+        of regularization, that regularization may bias coefficients,
+        improving the conditioning of the problem but harming the
+        quality of the fit. Setting ``unbias==True`` enables an extra
+        step wherein unregularized linear regression is applied, but
+        only for the coefficients in the support identified by the
+        optimizer. This helps to remove the bias introduced by
+        regularization.
+
     Attributes
     ----------
     coef_ : array, shape (n_features,) or (n_targets, n_features)
@@ -88,8 +99,9 @@ class BaseOptimizer(LinearRegression, ComplexityMixin):
         fit_intercept=False,
         initial_guess=None,
         copy_X=True,
+        unbias: bool = True,
     ):
-        super(BaseOptimizer, self).__init__(fit_intercept=fit_intercept, copy_X=copy_X)
+        super().__init__(fit_intercept=fit_intercept, copy_X=copy_X)
 
         if max_iter <= 0:
             raise ValueError("max_iter must be positive")
@@ -100,6 +112,7 @@ class BaseOptimizer(LinearRegression, ComplexityMixin):
             initial_guess = initial_guess.reshape(1, -1)
         self.initial_guess = initial_guess
         self.normalize_columns = normalize_columns
+        self.unbias = unbias
 
     # Force subclasses to implement this
     @abc.abstractmethod
@@ -178,6 +191,11 @@ class BaseOptimizer(LinearRegression, ComplexityMixin):
         self._reduce(x_normed, y, **reduce_kws)
         self.ind_ = np.abs(self.coef_) > 1e-14
 
+        self.history_.append(self.coef_)
+        if self.unbias:
+            self._unbias(x_normed, y)
+        self.history_.append(self.coef_)
+
         # Rescale coefficients to original units
         if self.normalize_columns:
             self.coef_ = np.multiply(reg, self.coef_)
@@ -188,6 +206,24 @@ class BaseOptimizer(LinearRegression, ComplexityMixin):
 
         self._set_intercept(X_offset, y_offset, X_scale)
         return self
+
+    def _unbias(self, x, y):
+        coef = np.zeros((y.shape[1], x.shape[1]))
+        if hasattr(self, "fit_intercept"):
+            fit_intercept = self.fit_intercept
+        else:
+            fit_intercept = False
+        for i in range(self.ind_.shape[0]):
+            if np.any(self.ind_[i]):
+                coef[i, self.ind_[i]] = (
+                    LinearRegression(fit_intercept=fit_intercept)
+                    .fit(x[:, self.ind_[i]], y[:, i])
+                    .coef_
+                )
+        if self.coef_.ndim == 1:
+            self.coef_ = coef[0]
+        else:
+            self.coef_ = coef
 
 
 class EnsembleOptimizer(BaseOptimizer):
@@ -249,12 +285,6 @@ class EnsembleOptimizer(BaseOptimizer):
     coef_full_ : array, shape (n_features,) or (n_targets, n_features)
         Weight vector(s) that are not subjected to the regularization.
         This is the w in the objective function.
-
-    unbias : boolean
-        Whether to perform an extra step of unregularized linear regression
-        to unbias the coefficients for the identified support.
-        ``unbias`` is automatically set to False if a constraint is used and
-        is otherwise left uninitialized.
     """
 
     def __init__(
@@ -271,7 +301,7 @@ class EnsembleOptimizer(BaseOptimizer):
         if not hasattr(opt, "initial_guess"):
             opt.initial_guess = None
 
-        super(EnsembleOptimizer, self).__init__(
+        super().__init__(
             max_iter=opt.max_iter,
             fit_intercept=opt.fit_intercept,
             initial_guess=opt.initial_guess,
@@ -302,6 +332,7 @@ class EnsembleOptimizer(BaseOptimizer):
         self.replace = replace
         self.n_candidates_to_drop = n_candidates_to_drop
         self.coef_list = []
+        self.unbias = False
 
     def _reduce(self, x: AxesArray, y: np.ndarray) -> None:
         x = AxesArray(np.asarray(x), {"ax_sample": 0, "ax_coord": 1})
